Add a faster git-breakout implementation
This implementation works without rebasing, completely isolated from
the current working tree, branch and index, which makes it more
comfortable and faster, at the cost of being unable to handle any
conflicts. But usually, that shouldn't be a problem for the kind of
commits that this command is made for anyway. And if there's a problem,
there's still the rebase version as a fallback.